Rename transformer_embeddings to sentence_transformer_embeddings and hf_embeddings to huggingface_embeddings
