+++ pyproject.toml
@@ -55,12 +55,12 @@ _metadata = "asreviewcontrib.dory._metadata:MODEL_METADATA"
 "xgboost" = "asreviewcontrib.dory.classifiers.xgboost:XGBoost"
 
 [project.entry-points."asreview.models.feature_extractors"]
-"sbert" = "asreviewcontrib.dory.feature_extractors.transformer_embeddings:SBERT"
-"mxbai" = "asreviewcontrib.dory.feature_extractors.transformer_embeddings:MXBAI"
-"labse" = "asreviewcontrib.dory.feature_extractors.transformer_embeddings:LaBSE"
-"multilingual-e5-large" = "asreviewcontrib.dory.feature_extractors.transformer_embeddings:MultilingualE5Large"
-"gtr-t5-large" = "asreviewcontrib.dory.feature_extractors.transformer_embeddings:GTR"
-"xlm-roberta-large" = "asreviewcontrib.dory.feature_extractors.hf_embeddings:XLMRoBERTaLarge"
+"sbert" = "asreviewcontrib.dory.feature_extractors.sentence_transformer_embeddings:SBERT"
+"mxbai" = "asreviewcontrib.dory.feature_extractors.sentence_transformer_embeddings:MXBAI"
+"labse" = "asreviewcontrib.dory.feature_extractors.sentence_transformer_embeddings:LaBSE"
+"multilingual-e5-large" = "asreviewcontrib.dory.feature_extractors.sentence_transformer_embeddings:MultilingualE5Large"
+"gtr-t5-large" = "asreviewcontrib.dory.feature_extractors.sentence_transformer_embeddings:GTR"
+"xlm-roberta-large" = "asreviewcontrib.dory.feature_extractors.huggingface_embeddings:XLMRoBERTaLarge"
 
 [project.entry-points."asreview.entry_points"]
 dory = "asreviewcontrib.dory.entrypoint:DoryEntryPoint"
